requirements.txt
torch 
wandb 
numpy
#pandas
#gensim
#mypy
#ruff
#black
#datasets
#pyarrow
#tqdm
#huggingface_hub
#faiss-cpu
#fastapi
#uvicorn
#torchvision
#torchaudio 
#matplotlib
#seaborn
#gradio
transformers
#pillow
#git+https://github.com/soundata/soundata.git@main
#essentia
#librosa
datasets
#git+https://github.com/openai/whisper.git
#openai
dotenv
#spacy
accelerate
peft 
bitsandbytes
accelerate 
trl
pillow
torchvision